Moved CreditCard class to the CreditCard.cs file
