+++ HelloWorld/HelloWorld/Phone.cs
@@ -67,41 +67,4 @@ namespace TMS_Hometasks
         {
         }
     }
-
-    //<summary>
-    //2. Создать класс CreditCard c полями номер счета, текущая сумма на счету.
-    //Добавьте метод, который позволяет начислять сумму на кредитную карточку.
-    //Добавьте метод, который позволяет снимать с карточки некоторую сумму.
-    //Добавьте метод, который выводит текущую информацию о карточке.
-    //Напишите программу, которая создает три объекта класса CreditCard у которых заданы номер счета и начальная сумма
-    //Тестовый сценарий для проверки:
-    //Положите деньги на первые две карточки и снимите с третьей.
-    //Выведите на экран текущее состояние всех трех карточек.
-    //</summary>
-    internal class CreditCard
-    {
-        public int cardNumber;
-        public double cardBalance;
-
-        public double AddSum(double sum)
-        {
-            return cardBalance += sum;
-        }
-
-        public double RemoveSum(double sum)
-        {
-            return cardBalance -= sum;
-        }
-
-        public void GetInfo()
-        {
-            Console.WriteLine($"Card number is {cardNumber}, balance on the card is {cardBalance}.");
-        }
-
-        public CreditCard(int cardNumber, double cardBalance)
-        {
-            this.cardNumber = cardNumber;
-            this.cardBalance = cardBalance;
-        }
-    }
 }
